requirements.txt
numpy==1.26.4
pyaudio==0.2.14
pyttsx3==2.90
librosa==0.10.2
torch==2.3.1
pyannote-audio==3.1.1
openai-whisper